+++ src/main/java/pl/agilevision/hardware/um7/data/GyroProcessed.java
@@ -11,7 +11,8 @@ public class GyroProcessed extends BaseAttribute {
   public static String Z = "gyro_proc_z";
   public static String Time = "gyro_proc_time";
 
-  public GyroProcessed(long registerAddress, int bitOffset) {
+
+  public GyroProcessed(int registerAddress, int bitOffset) {
     super(registerAddress, bitOffset);
   }
 }
